pyproject.toml
[tool]
[tool.pdm]

[project]
name = "Aniblog"
version = "1.0.0"
description = ""
authors = [
    {name = "Dzmitry Zhybryk", email = "mr.zhybryk@gmail.com"},
]
dependencies = [
    "grpcio>=1.51.1",
    "grpcio-tools>=1.51.1",
    "pydantic[dotenv]>=1.10.5",
    "sqlalchemy>=2.0.3",
    "python-jose>=3.3.0",
    "psycopg[binary]>=3.1.8",
]
requires-python = ">=3.10"
license = {text = "MIT"}
